Updated Client
Moved company list api from crew to client
